Extract Xtext type editor page
Extract Xtext type editor page to be reused outside of the FB type
editor.
