+++ plugins/org.eclipse.fordiac.ide.fbtypeeditor/plugin.properties
@@ -21,9 +21,6 @@ FBInterfacePaletteFactory_EventTypes=Event types
 FBInterfacePaletteFactory_DataTypes=Data types
 FBInterfacePaletteFactory_RootGroup=Root Group
 FBTypeEditor = FB Type Editor
-FBTypeXtextEditor_PerformanceModeDialogMessage=Enabled performance mode since the current source exceeds the configured line limit. No code minings, folding, semantic hightlighting, or outline available.
-FBTypeXtextEditor_PerformanceModeDialogTitle=Performance Mode
-FBTypeXtextEditor_PerformanceModeOutlineMessage=No Outline available in performance mode
 
 FunctionBlockTypeEditor.bundleName = Function Block Type Editor
