+++ plugins/org.eclipse.fordiac.ide.typeeditor/META-INF/MANIFEST.MF
@@ -8,7 +8,8 @@ Bundle-Vendor: Eclipse 4diac
 Automatic-Module-Name: org.eclipse.fordiac.ide.typeeditor
 Bundle-RequiredExecutionEnvironment: JavaSE-21
 Bundle-ActivationPolicy: lazy
-Import-Package: org.eclipse.core.commands.common,
+Import-Package: com.google.inject,
+ org.eclipse.core.commands.common,
  org.eclipse.core.expressions,
  org.eclipse.core.resources,
  org.eclipse.core.runtime,
@@ -35,6 +36,8 @@ Import-Package: org.eclipse.core.commands.common,
  org.eclipse.jface.dialogs,
  org.eclipse.jface.layout,
  org.eclipse.jface.resource,
+ org.eclipse.jface.text.source,
+ org.eclipse.jface.text.source.projection,
  org.eclipse.jface.viewers,
  org.eclipse.ocl.internal.l10n,
  org.eclipse.osgi.util,
@@ -48,8 +51,11 @@ Import-Package: org.eclipse.core.commands.common,
  org.eclipse.ui.forms.editor,
  org.eclipse.ui.ide,
  org.eclipse.ui.part,
+ org.eclipse.ui.plugin,
  org.eclipse.ui.texteditor,
  org.eclipse.ui.views.contentoutline,
  org.eclipse.ui.views.properties,
  org.eclipse.ui.views.properties.tabbed,
- org.eclipse.xtext.ui.editor
+ org.eclipse.xtext.resource,
+ org.eclipse.xtext.ui.editor,
+ org.osgi.framework
