+++ plugins/org.eclipse.fordiac.ide.typeeditor/src/org/eclipse/fordiac/ide/typeeditor/messages.properties
@@ -22,3 +22,6 @@ TypeEditor_TypeUpdateDialog_Description=The following types were affected by cha
 TypeEditor_WarningDialog_Headline=Type Editor
 TypeEditor_WarningDialog_NotSaved=The changes have not been saved!
 
+XtextTypeEditorPage_PerformanceModeDialogMessage=Enabled performance mode since the current source exceeds the configured line limit. No code minings, folding, semantic hightlighting, or outline available.
+XtextTypeEditorPage_PerformanceModeDialogTitle=Performance Mode
+XtextTypeEditorPage_PerformanceModeOutlineMessage=No Outline available in performance mode
